# openapi.json
{
  "components": {
    "schemas": {
...
      "AnalyzeIn": {
        "description": "POST /analysis — analyse a target for a card.",
        "properties": {
          "kind": {
            "enum": [
              "track",
              "remains"
            ],
            "title": "Kind",
            "type": "string"
          },
          "threat_id": {
            "title": "Threat Id",
            "type": "integer"
          }
        },
        "required": [
          "threat_id",
          "kind"
        ],
        "title": "AnalyzeIn",
        "type": "object"
      },
      "AnalyzeOut": {
        "description": "A successful analysis: the card that dropped.",
        "properties": {
          "card_id": {
            "title": "Card Id",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "kind": {
            "enum": [
              "track",
              "remains"
            ],
            "title": "Kind",
            "type": "string"
          },
          "threat_id": {
            "title": "Threat Id",
            "type": "integer"
          }
        },
        "required": [
          "threat_id",
          "kind",
          "card_id",
          "created_at"
        ],
        "title": "AnalyzeOut",
        "type": "object"
      },
...
      "CardCountOut": {
        "description": "One collected card + how many copies the user has.",
        "properties": {
          "card_id": {
            "title": "Card Id",
            "type": "integer"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "first_at": {
            "format": "date-time",
            "title": "First At",
            "type": "string"
          }
        },
        "required": [
          "card_id",
          "count",
          "first_at"
        ],
        "title": "CardCountOut",
        "type": "object"
      },
      "CollectionOut": {
        "description": "GET /analysis/collection — the current user's whole card collection.",
        "properties": {
          "card_count": {
            "title": "Card Count",
            "type": "integer"
          },
          "cards": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/CardCountOut"
            },
            "title": "Cards",
            "type": "array"
          },
          "total_analyses": {
            "title": "Total Analyses",
            "type": "integer"
          }
        },
        "required": [
          "total_analyses",
          "card_count"
        ],
        "title": "CollectionOut",
        "type": "object"
      },
...
      "GamificationPrefIn": {
        "description": "PUT /me/gamification — flip the account-bound gamification toggle.",
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "enabled"
        ],
        "title": "GamificationPrefIn",
        "type": "object"
      },
      "GamificationPrefOut": {
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "enabled"
        ],
        "title": "GamificationPrefOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ThreatAnalysisStateOut": {
        "description": "GET /analysis/threat/{id} — which analyses this target has left, and which\n(if any) the current user already claimed. Drives the inspect-badge button:\n`*_taken` disables it globally, `mine.*` shows the card the user won.",
        "properties": {
          "mine_remains": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mine Remains"
          },
          "mine_track": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mine Track"
          },
          "remains_taken": {
            "title": "Remains Taken",
            "type": "boolean"
          },
          "track_taken": {
            "title": "Track Taken",
            "type": "boolean"
          }
        },
        "required": [
          "track_taken",
          "remains_taken"
        ],
        "title": "ThreatAnalysisStateOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Kyiv Aerial Threat Tracker (MVP)",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/analysis": {
      "post": {
        "operationId": "analyze_target_analysis_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyze Target",
        "tags": [
          "gamification"
        ]
      }
    },
    "/analysis/collection": {
      "get": {
        "operationId": "my_collection_analysis_collection_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "My Collection",
        "tags": [
          "gamification"
        ]
      }
    },
    "/analysis/threat/{threat_id}": {
      "get": {
        "operationId": "threat_analysis_state_analysis_threat__threat_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "threat_id",
            "required": true,
            "schema": {
              "title": "Threat Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreatAnalysisStateOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Threat Analysis State",
        "tags": [
          "gamification"
        ]
      }
    },
...
    "/collection/{user_id}": {
      "get": {
        "description": "Another user's collection — visible only to that user themselves or an\naccepted friend (collections aren't public).",
        "operationId": "friend_collection_collection__user_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Friend Collection",
        "tags": [
          "gamification"
        ]
      }
    },
...
    "/me/gamification": {
      "put": {
        "description": "Persist the account-bound gamification toggle so it syncs across devices.",
        "operationId": "set_gamification_pref_me_gamification_put",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GamificationPrefIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GamificationPrefOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Gamification Pref",
        "tags": [
          "gamification"
        ]
      }
    },
...
  }
}